Azure

+++ scripts/undeploy.sh
@@ -9,8 +9,8 @@ default_helper() {
         kubefs undeploy --help - display this help message
 
         Args:
-            --target | -t <local|EKS|Azure|GCP> - specify the deployment target for which cluster (default is local)
-            --close-minikube | -cm - close minikube after undeploying components
+            --target | -t <local|EKS|azure|GCP> - specify the deployment target for which cluster (default is local)
+            --close | -c - stop cluster after undeploying components
     "
 }
 
@@ -18,24 +18,46 @@ parse_optional_params(){
     declare -A opts
 
     opts["--target"]=local
-    opts["--close-minikube"]=false
+    opts["--close"]=false
 
     while [ $# -gt 0 ]; do
         case $1 in
             --target | -t)
-                if [ "$2" == "local" ] || [ "$2" == "EKS" ] || [ "$2" == "Azure" ] || [ "$2" == "GCP" ]; then
+                if [ "$2" == "local" ] || [ "$2" == "EKS" ] || [ "$2" == "azure" ] || [ "$2" == "GCP" ]; then
                     opts["--target"]=$2
                     shift
                 fi 
                 ;;
-            --close-minikube | -cm)
-                opts["--close-minikube"]=true
+            --close | -c)
+                opts["--close"]=true
                 ;;
         esac
         shift
     done
 
     echo $(declare -p opts)
+}
+
+stop(){
+    TARGET=$1
+    case $TARGET in
+        # "EKS") ;;
+        "azure")
+            az aks stop --name $(yq e '.azure.cluster_name' $CURRENT_DIR/manifest.yaml) --resource-group $(yq e '.azure.resource_group' $CURRENT_DIR/manifest.yaml)
+            if [ $? -eq 1 ]; then
+                print_error "Error occured stopping the cluster. Please try again or use 'kubefs --help' for more information."
+                return 1
+            fi 
+        ;;
+        # "GCP") ;;
+        *)
+            minikube stop
+            if [ $? -eq 1 ]; then
+                print_error "Error occured stopping the cluster. Please try again or use 'kubefs --help' for more information."
+                return 1
+            fi
+        ;;
+    esac
 }
 
 undeploy_all(){
@@ -57,11 +79,12 @@ undeploy_all(){
         fi
     done
 
-    if [ "${opts["--close-minikube"]}" == true ]; then
+    if [ ${opts["--close"]} = true ]; then
         while kubectl get namespaces | grep -q terminating; do
             sleep 2
         done
-        minikube stop
+        
+        stop ${opts["--target"]}
     fi
 
     print_success "Undeployed all components"
@@ -80,13 +103,39 @@ undeploy_helper(){
         return 0
     fi
 
-    if [ "${opts["--close-minikube"]}" == true ]; then
+    if [ ${opts["--close"]} = true ]; then
         while kubectl get namespaces | grep -q terminating; do
-            sleep 5
+            sleep 2
         done
-        minikube stop
+        
+        stop ${opts["--target"]}
     fi
 
+    return 0
+}
+
+undeploy_azure(){
+    NAME=$1
+    echo "Undeploying $NAME from Azure..."
+
+    if ! az account show > /dev/null 2>&1; then
+        print_warning "Azure account not logged in. Please login using 'kubefs config azure'"
+        return 1
+    fi
+
+    az aks get-credentials --name $(yq e '.azure.cluster_name' $CURRENT_DIR/manifest.yaml) --resource-group $(yq e '.azure.resource_group' $CURRENT_DIR/manifest.yaml)
+    if [ $? -eq 1 ]; then
+        print_error "Error occured deploying $NAME. Please try again or use 'kubefs --help' for more information."
+        return 1
+    fi
+
+    helm uninstall $NAME     
+    if [ $? -eq 1 ]; then
+        print_error "Error occured deploying $NAME. Please try again or use 'kubefs --help' for more information."
+        return 1
+    fi
+
+    print_success "$NAME undeployed successfully"
     return 0
 }
 
@@ -110,11 +159,16 @@ undeploy_unique(){
     
     case ${opts["--target"]} in
         # "EKS") deploy_eks $NAME;;
-        # "Azure") deploy_azure $NAME;;
+        "azure") undeploy_azure $NAME;;
         # "GCP") deploy_gcp $NAME;;
         *)
-            helm uninstall $NAME 
+            kubectl config use-context minikube
+            if [ $? -eq 1 ]; then
+                print_error "Error occured deploying $NAME. Please try again or use 'kubefs --help' for more information."
+                return 1
+            fi
 
+            helm uninstall $NAME 
             if [ $? -eq 1 ]; then
                 print_error "Error occured deploying $NAME. Please try again or use 'kubefs --help' for more information."
                 return 1
